+++ video-wallpaper/settings/GeneralTab.qml
@@ -19,11 +19,19 @@ ColumnLayout {
     readonly property bool isPlaying: 
         pluginApi.pluginSettings.isPlaying ||
         false
-    
+
 
     property string currentWallpaper: 
         pluginApi?.pluginSettings?.currentWallpaper || 
         ""
+
+    property int fillMode:
+        pluginApi?.pluginSettings?.fillMode ||
+        0
+
+    property int orientation:
+        pluginApi?.pluginSettings?.orientation ||
+        0
 
     property double volume:
         pluginApi?.pluginSettings?.volume ||
@@ -114,15 +122,74 @@ ColumnLayout {
 
     NDivider {}
 
+    // Fill Mode
+    NComboBox {
+        enabled: root.enabled
+        Layout.fillWidth: true
+        label: root.pluginApi?.tr("settings.general.fill_mode.label") || "Fill Mode"
+        description: root.pluginApi?.tr("settings.general.fill_mode.description") || "The mode that the wallpaper is fitted into the background."
+        defaultValue: "0"
+        model: [
+            {
+                "key": "0",
+                "name": root.pluginApi?.tr("settings.general.fill_mode.stretch") || "Stretch"
+            },
+            {
+                "key": "1",
+                "name": root.pluginApi?.tr("settings.general.fill_mode.fit") || "Fit"
+            },
+            {
+                "key": "2",
+                "name": root.pluginApi?.tr("settings.general.fill_mode.crop") || "Crop"
+            }
+        ]
+        currentKey: root.fillMode
+        onSelected: key => root.fillMode = key
+    }
+
+    // Orientation
+    NValueSlider {
+        property real _value: root.orientation
+
+        enabled: root.enabled
+        from: -270
+        to: 270
+        value: root.orientation
+        defaultValue: 0
+        stepSize: 90
+        text: _value
+        label: root.pluginApi?.tr("settings.general.orientation.label") || "Orientation"
+        description: root.pluginApi?.tr("settings.general.orientation.description") || "The orientation of the video playing, can be any multiple of 90."
+        onMoved: value => _value = value
+        onPressedChanged: (pressed, value) => {
+            if(root.pluginApi == null) {
+                Logger.e("video-wallpaper", "Plugin API is null.");
+                return
+            }
+
+            if(!pressed) {
+                root.pluginApi.pluginSettings.orientation = value;
+                root.pluginApi.saveSettings();
+            }
+        }
+    }
+
+    NDivider {}
+
     // Volume
     NValueSlider {
+        property real _value: root.volume
+
         enabled: root.enabled
-        from: 0
-        to: 1
+        from: 0.0
+        to: 1.0
+        defaultValue: 1.0
         value: root.volume
         stepSize: (Settings.data.audio.volumeStep / 100.0)
+        text: `${_value * 100.0}%`
         label: root.pluginApi?.tr("settings.general.volume.label") || "Volume"
         description: root.pluginApi?.tr("settings.general.volume.description") || "The current volume of the video playing."
+        onMoved: value => _value = value
         onPressedChanged: (pressed, value) => {
             if(root.pluginApi == null) {
                 Logger.e("video-wallpaper", "Plugin API is null.");
@@ -156,6 +223,8 @@ ColumnLayout {
         }
 
         pluginApi.pluginSettings.currentWallpaper = currentWallpaper;
+        pluginApi.pluginSettings.orientation = orientation;
+        pluginApi.pluginSettings.fillMode = fillMode;
         pluginApi.pluginSettings.wallpapersFolder = wallpapersFolder;
     }
 }